AWS DB 연결, DB오류들 수정
최신 repo에서 끌어오고 다시 작업함
