+++ .gitignore
@@ -28,3 +28,11 @@ hs_err_pid*
 "comments/\353\257\270\354\271\234\353\206\210.txt"
 "comments/\354\202\254\353\236\221\355\225\264.txt"
 "comments/\354\204\234\353\251\264\354\227\255\354\227\220\354\204\234.txt"
+
+~$해결기법 제안서.docx
+
+~WRL0005.tmp
+
+out/production/MusicChartWithCommunity/문제해결기법 제안서.docx
+
+out/production/MusicChartWithCommunity/comments/*.txt
