+++ .idea/dataSources/511465f6-5e69-4892-a4eb-b9f1271d8e84/storage_v2/_src_/schema/information_schema.FNRwLQ.meta
@@ -1,0 +1,2 @@
+#n:information_schema
+!<md> [null, 0, null, null, -2147483648, -2147483648]
