+++ src/DB/ConnectDB.java
@@ -5,12 +5,11 @@ import java.sql.*;
 public class ConnectDB {
     public static Connection GetDB() {
         Connection con = null;
-        PreparedStatement pstmt = null;
-        Statement stmt = null;
-        ResultSet rs;
-        String url = "jdbc:mysql://localhost:3306/charts_db?&serverTimezone=Asia/Seoul&useSSL=false";
-        String userid = "root";
-        String pwd = "root";
+
+        String url = "jdbc:mysql://database-2.cqcrpm8zqs1t.ap-northeast-2.rds.amazonaws.com:3306/charts_db?&serverTimezone=Asia/Seoul&useSSL=false";
+        String userid = "admin";
+        String pwd = "refactoring";
+
         // 1.드라이버 로딩
         try {
             Class.forName("com.mysql.jdbc.Driver");
@@ -24,30 +23,6 @@ public class ConnectDB {
             System.out.println("데이터베이스 연결 준비...");
             con = DriverManager.getConnection(url, userid, pwd);
             System.out.println("데이터베이스 연결 성공");
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-
-        try{
-            stmt = con.createStatement();
-            String sql = "SELECT table_name FROM information_schema.tables where table_schema = 'charts_db' and table_name = 'songinfo'";
-            rs = stmt.executeQuery(sql);
-
-            if(!rs.next()){
-                System.out.println("데이터베이스 테이블 생성");
-                //테이블이 없다면 생성
-
-                sql = "create table songinfo ("
-                        + "`title` varchar(100) COLLATE utf8_bin NOT NULL,"
-                        + "`artist` varchar(100) COLLATE utf8_bin NOT NULL,"
-                        + "`albumname` varchar(100) COLLATE utf8_bin NOT NULL,"
-                        + "`siteid` int NOT NULL,"
-                        + "`comment` varchar(100) COLLATE utf8_bin NOT NULL,"
-                        + "`pwd` varchar(8) COLLATE utf8_bin NOT NULL"
-                        + ")";
-                boolean rs2 = stmt.execute(sql);
-                stmt.close();
-            }
         } catch (SQLException e) {
             e.printStackTrace();
         }
