+++ src/view/SitePanel.java
@@ -1,6 +1,7 @@
 package view;
 
 import java.awt.*;
+import java.awt.event.KeyListener;
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
 import java.net.MalformedURLException;
@@ -10,7 +11,7 @@ import javax.swing.*;
 import javax.swing.table.*;
 
 import notsort.AppManager;
-import notsort.MusicChartParser;
+import model.MusicChartParser;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 
@@ -25,10 +26,10 @@ public class SitePanel extends JPanel {
 	private String strChartName;
 	
 	//음악 차트를 담는 표
-	private JTable tableChart;
+	public JTable tableChart;
 	
 	//표의 모델(셀의 크기, 개수, 표시 자료형 등을 결정)
-	private ChartModel tableModel;
+	public ChartModel tableModel;
 	
 	//표에서 정렬 및 필터링 기능을 담당
 	private TableRowSorter<ChartModel> tableSorter;
@@ -267,7 +268,11 @@ public class SitePanel extends JPanel {
 			return false;
 		}
 	} //ChartModel 클래스 끝
-	
+
+	public void addClickListener(MouseListener listenForMouse) {
+		tableChart.addMouseListener((listenForMouse));
+	}
+
 	// - - - - - ClickListener 클래스 - - - - -
 	private class ClickListener implements MouseListener {
 		@Override
@@ -278,10 +283,12 @@ public class SitePanel extends JPanel {
 		public void mouseClicked(MouseEvent e) {
 			Object obj = e.getSource();
 			if(obj == tableChart) {
+				/*
 				JTable table = (JTable) obj;
 				Object[] music = tableModel.getMusicData(table.convertRowIndexToModel(table.getSelectedRow())); //클릭된 열의 위치(숨겨진 항목이 있어도 바뀌지 않는 절대적인 위치)에 있는 곡 선택
 				System.out.println(music[2] + music[0].toString()); //테스트
 				AppManager.getS_instance().PopUpCommentUI(Integer.parseInt(music[0].toString())); //선택된 곡에 대한 커뮤니티 표시
+				*/
 			}
 		}
 		@Override
